fix: Dark mode button remove border in button design it responsive
