+++ src/components/ProgressStepper/ProgressStepper.scss
@@ -1,6 +1,4 @@
 
-/// 
-//src/components/ProgressStepper/ProgressStepper.scss
 @use "sass:color";
 @use "../../styles/variables" as *;
