+++ src/components/Sidebar/sidebar.scss
@@ -1,17 +1,16 @@
 
-//src/components/Sidebar/Sidebar.scss
 @use "sass:color";
 @use "../../styles/variables" as *;
 
 .sidebar {
-  position: fixed; // Changed from relative to fixed
+  position: fixed;
   z-index: 10;
   transition: all 0.3s ease-in-out;
   flex-shrink: 0;
   height: 100vh;
   background-color: var(--paper-background);
   border-right: 1px solid var(--border-color);
-  overflow-x: hidden; // Prevent horizontal scroll
+  overflow-x: hidden;
   
   &.open {
     width: 256px;
@@ -43,8 +42,28 @@
       cursor: pointer;
       transition: all 0.2s ease-in-out;
       z-index: 20;
+      outline: none;  // Added this
       
       &:hover {
+        transform: translateY(-50%) scale(1.1);
+        background-color: var(--hover-bg);
+      }
+    
+      // Add these new styles
+      &:focus {
+        outline: none;
+      }
+    
+      &:focus-visible {
+        outline: none;
+      }
+    
+      &::-moz-focus-inner {
+        border: 0;
+      }
+    
+      // Optional: if you want to keep the hover effect when focused
+      &:focus:hover {
         transform: translateY(-50%) scale(1.1);
         background-color: var(--hover-bg);
       }
@@ -64,10 +83,10 @@
       flex-grow: 1;
       display: flex;
       flex-direction: column;
-      overflow-y: auto; // Enable scroll for navigation
-      overflow-x: hidden; // Prevent horizontal scroll
-      margin-right: -#{$spacing-md}; // Compensate for padding
-      padding-right: $spacing-md; // Add padding back
+      overflow-y: auto;
+      overflow-x: hidden;
+      margin-right: -#{$spacing-md};
+      padding-right: $spacing-md;
 
       .nav-item-wrapper {
         margin-bottom: $spacing-sm;
@@ -84,22 +103,39 @@
         cursor: pointer;
         text-decoration: none;
         white-space: nowrap;
-
+        width: 100%;
+        color: inherit;
+        background: transparent;
+        border: none;
+        outline: none;  // Added this
+      
         &:hover {
           background-color: var(--hover-bg);
         }
-
+      
+        &:focus {
+          outline: none;
+        }
+      
+        &:focus-visible {
+          outline: none;
+        }
+      
+        &::-moz-focus-inner {
+          border: 0;
+        }
+      
         .item-icon {
           width: 20px;
           height: 20px;
           min-width: 20px;
         }
-
+      
         .item-text {
           margin-left: $spacing-md;
           transition: color 0.2s ease-in-out;
         }
-
+      
         .arrow-icon {
           margin-left: auto;
           width: 16px;
@@ -125,7 +161,6 @@
         }
       }
 
-      // Scrollbar styling for navigation
       &::-webkit-scrollbar {
         width: 6px;
       }
@@ -141,34 +176,6 @@
         &:hover {
           background: var(--text-color);
         }
-      }
-    }
-
-    .theme-toggle {
-      padding: $spacing-md;
-      border-radius: 0.5rem;
-      cursor: pointer;
-      transition: background-color 0.2s ease-in-out;
-      display: flex;
-      align-items: center;
-      background-color: transparent;
-      margin-top: $spacing-md;
-      border: none;
-
-      &:hover {
-        background-color: var(--hover-bg);
-      }
-
-      .theme-icon {
-        width: 20px;
-        height: 20px;
-        color: var(--primary-color);
-      }
-
-      .theme-text {
-        margin-left: $spacing-md;
-        color: var(--text-color);
-        font-size: 0.875rem;
       }
     }
   }
