+++ src/context/ThemeContext.tsx
@@ -1,39 +1,11 @@
 // src/context/ThemeContext.tsx
-import React, { createContext, useContext, useState, useCallback, useEffect } from 'react';
+import React from 'react';
 import { ThemeProvider as MuiThemeProvider } from '@mui/material/styles';
 import { createAppTheme } from '../theme/theme';
-
-interface ThemeContextType {
-  mode: 'light' | 'dark';
-  toggleTheme: () => void;
-}
-
-const ThemeContext = createContext<ThemeContextType>({
-  mode: 'light',
-  toggleTheme: () => {},
-});
-
-export const useTheme = () => useContext(ThemeContext);
+import { useThemeLogic, ThemeContext } from './ThemeContextLogic';
 
 export const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-  const [mode, setMode] = useState<'light' | 'dark'>(() => {
-    // Get saved theme from localStorage or use system preference
-    const savedTheme = localStorage.getItem('theme');
-    if (savedTheme === 'dark' || savedTheme === 'light') {
-      return savedTheme;
-    }
-    return window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light';
-  });
-
-  // Update the HTML data-theme attribute and localStorage when theme changes
-  useEffect(() => {
-    document.documentElement.setAttribute('data-theme', mode);
-    localStorage.setItem('theme', mode);
-  }, [mode]);
-
-  const toggleTheme = useCallback(() => {
-    setMode((prevMode) => (prevMode === 'light' ? 'dark' : 'light'));
-  }, []);
+  const { mode, toggleTheme } = useThemeLogic();
 
   // Create theme object (assuming you're using Material-UI)
   const theme = React.useMemo(() => createAppTheme(mode), [mode]);
