+++ src/context/ThemeContextLogic.ts
@@ -1,0 +1,36 @@
+// src/context/ThemeContextLogic.ts
+import React, { createContext, useState, useCallback, useEffect } from 'react';
+
+interface ThemeContextType {
+  mode: 'light' | 'dark';
+  toggleTheme: () => void;
+}
+
+// Export ThemeContext
+export const ThemeContext = createContext<ThemeContextType>({
+  mode: 'light',
+  toggleTheme: () => {},
+});
+
+export const useThemeLogic = () => {
+  const [mode, setMode] = useState<'light' | 'dark'>(() => {
+    // Get saved theme from localStorage or use system preference
+    const savedTheme = localStorage.getItem('theme');
+    if (savedTheme === 'dark' || savedTheme === 'light') {
+      return savedTheme;
+    }
+    return window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light';
+  });
+
+  // Update the HTML data-theme attribute and localStorage when theme changes
+  useEffect(() => {
+    document.documentElement.setAttribute('data-theme', mode);
+    localStorage.setItem('theme', mode);
+  }, [mode]);
+
+  const toggleTheme = useCallback(() => {
+    setMode((prevMode) => (prevMode === 'light' ? 'dark' : 'light'));
+  }, []);
+
+  return { mode, toggleTheme };
+};
