+++ src/context/useTheme.ts
@@ -1,0 +1,5 @@
+// src/context/useTheme.ts
+import { useContext } from 'react';
+import { ThemeContext } from './ThemeContextLogic';
+
+export const useTheme = () => useContext(ThemeContext);
